feat: upgrade register page style
