+++ Cardoon/src/components/pages/RegisterPage/RegisterPage.scss
@@ -1,0 +1,26 @@
+@use "../../../styles/variables.scss" as *;
+
+.RegisterPage {
+  height: $PAGE_HEIGHT;
+  width: $PAGE_WIDTH;
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  justify-content: center;
+
+  background: $background-gradient;
+
+  &__form {
+    height: auto;
+    max-width: 400px;
+    padding: 20px;
+    border-radius: 8px;
+    box-shadow: 0 2px 10px rgba(0, 0, 0, 0.1);
+    background-color: white;
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+    justify-content: center;
+    gap: 8px;
+  }
+}
