+++ Cardoon/src/styles/app.scss
@@ -9,6 +9,7 @@
 @use "../components/pages/UpgradePage/UpgradePage" as UpgradePage;
 @use "../components/pages/BossPage/BossPage" as BossPage;
 @use "../components/pages/GamePage/GamePageStyle" as GamePage;
+@use "../components/pages/RegisterPage/RegisterPage" as RegisterPage;
 @use "../components/pages/TokenErrorPage/TokenErrorPage" as NotFoundPage;
 @use "../components/pages/CardFormPage/CardFormPage" as CardFormPage;
 @use "../components/molecules/EditCardForm/EditCardForm" as EditCardForm;
