+++ Cardoon/src/styles/variables.scss
@@ -14,3 +14,4 @@ $disabled-text-color: #9e9e9e;
 
 $NAVBAR_HEIGHT: 60px;
 $PAGE_HEIGHT: calc(100vh - #{$NAVBAR_HEIGHT});
+$PAGE_WIDTH: 100%;
